Add charge to fmwpc output

+++ src/GlueXSensitiveDetectorCDC.hh
@@ -58,7 +58,7 @@ class GlueXSensitiveDetectorCDC : public G4VSensitiveDetector
    static double CDC_TIME_WINDOW;
    static double GAS_GAIN;
    static double W_EFF_PER_ION;
-   static int N_SECOND_PER_PRIMARY;
+   static double N_SECOND_PER_PRIMARY;
    static int MAX_HITS;
 
    static int fDrift_clusters;
